requirements.txt
alembic==1.9.1
click==8.1.3
Flask==2.2.2
Flask-Migrate==4.0.0
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
importlib-metadata==5.2.0
importlib-resources==5.10.1
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.4
MarkupSafe==2.1.1
python-dotenv==0.21.0
SQLAlchemy==1.4.45
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.11.0
